@bnoushin7 issue 132 resolved

+++ development-codes/update.py
@@ -1,23 +1,7 @@
 import json
 import os
 import yaml
-
-
-
-def make_html(current, parent, ancestors):
-    
-    catalog_dir = "https://raw.githubusercontent.com/kpegion/COLA-DATASETS-CATALOG/gh-pages/intake-catalogs/"
-    open_catalog = catalog_dir + current +".yaml"
-    title = current
-
-    _header = src_header(title, ancestors,  open_catalog)
-
-    html_page = current +".html"
-
-    with open(html_page , "w", encoding='utf-8') as file:
-        file.write(_header)
-        print( html_page + " was created\n")
-
+import pdb
 
 
 
@@ -41,16 +25,24 @@ def update_json(tags, html_page, dataset_sub_name):
 
 
 
-def update_parents(ancestors):
+def make_ancestors(ancestors):
     #<!-- 1qaz2wsx -->
-    
     ancestors = ancestors.split('/')
-    ancestors.remove('')
+
+    try:
+        ancestors.remove('')
+    except:
+        pass
+
+    
+
     try:
         ancestors.remove('shared')
     except:
         pass
 
+    print(ancestors)
+    print("------------------------------------------------------------------")
 
     try:
         ancestors.remove('data')
@@ -58,9 +50,13 @@ def update_parents(ancestors):
     except:
         pass
 
+    #pdb.set_trace()
 
+    #print(ancestors)
+    #print("------------------------------------------------------------------")
     ancestors = list(map(lambda x:x.lower(),ancestors))
-
+    #print(ancestors)
+    #print("------------------------------------------------------------------")
 
     res = ""
     for i in range(len(ancestors)):
@@ -73,56 +69,64 @@ def update_parents(ancestors):
 
 
 
-
-def insert_links(direct_parent):
-    
-    dp_name = direct_parent+".html"
-    searchee = "<!--qazwsxxswzaq-->"
-    res = """               <a href="precip.pentad.mean.nc" class="list-group-item"> 
-                        <h4>cmap_pentad</h4> 
-                        <p class="description"> </p> 
-                    </a> """
-    res = res + "\n\n" + searchee
-    fp = open(dp_name, "rt")
-    fin = fp.read()
-    data = fin.replace(searchee, res)
-    fp.close()
-
-    fp = open(dp_name, "wt")
-    fp.write(data)
-    fp.close()      
-    
-
-
-
 def catalog_parent( _path, _dataset_name):
 
-    
+    print(_dataset_name) 
     direct_parent = _path.split('/')[-2].lower()
     dict_file = [
-    {'description': 'COLDA '+ direct_parent + ' Data Caalog'},
+    {'description': 'COLDA '+ direct_parent.upper() + ' Data Caalog'},
     {'sources':{_dataset_name:[{'args':{'path':_path}},{'description':''},{'driver':'intake.catalog.local.YAMLFileCatalog'},{'metadata':'{}'}]
     }}]
-    
     direct_parent = direct_parent + '.yaml'
-    with open(direct_parent, 'w') as fp:
+    with open(direct_parent, 'a') as fp:
         yaml.dump(dict_file, fp)
+        print(direct_parent + " created")
 
-def update_links(child, dp_name):
 
+
+def catalog_parent_2( _path, _dataset_name, direct_parent):
+
+    direct_parent = direct_parent + '.yaml'
+    if not os.path.isfile(direct_parent):
+     
+        
+        dict_file = [
+        {'description': 'COLDA '+ direct_parent.upper() + ' Data Caalog'},
+        {'sources':{_dataset_name:[{'args':{'path':_path}},{'description':''},{'driver':'intake.catalog.local.YAMLFileCatalog'},{'metadata':'{}'}]
+        }}]
+
+    
+        with open(direct_parent, 'w') as fp:
+            yaml.dump(dict_file, fp)
+            print(direct_parent + " created")
+        
+    else:
+        
+        with open(direct_parent) as fp:
+            newdct = yaml.load(fp, Loader=yaml.FullLoader)
+
+        with open(direct_parent, 'w') as fp:
+            newdct[1]['sources'][_dataset_name+"tesdgdghdghtak"] =[{'args':{'path':_path}},{'description':'JadidTar'},{'driver':'intake.catalog.local.YAMLFileCatalog'},{'metadata':'{}'}]
+            yaml.dump(newdct, fp)
+
+def link_to_children(child, dp_name):
+
+    
     searchee = "<!--qazwsxxswzaq-->"
-    res = """               <a href="""+ "\""+ child+"\""+""" class="list-group-item">
+    res = """<a href="""+ "\""+ child+"\""+""" class="list-group-item">
                     <h4>""" +child+ """</h4>
                     <p class="description"> </p>
                 </a> """
-
     
+    #pdb.set_trace() 
     file_name = dp_name + ".html"
-    res = res + "\n\n" + searchee
 
     fp = open('../'+file_name, "r+")
     fin = fp.read()
-    data = fin.replace(searchee, res)
-    fp.seek(0,0)
-    fp.write(data)
 
+    if (fin.find(res) == -1): 
+        res = res + "\n\n" + searchee
+        data = fin.replace(searchee, res)
+        fp.seek(0,0)
+        fp.write(data)
+    fp.close()
